created toolkit as a flyweight to eliminate cyclical dependencies
